[ALPS02335158] Cam: Modify IMX258 sensor driver
[Detail]Modify IMX258 PDAF setting

[Solution]Done

[Feature] [Android Default] Camera Application Basic Functions

MTK-Commit-Id: 47c8c704ab79757edb5af843eac82973c2c2ab9d

CR-Id: ALPS02335158

+++ drivers/misc/mediatek/imgsensor/src/mt6797/s5k5e2ya_mipi_raw/s5k5e2yamipiraw_Sensor.c
@@ -26,8 +26,7 @@
 #include <linux/uaccess.h>
 #include <linux/fs.h>
 #include <asm/atomic.h>
-//#include <asm/system.h>
-//#include <linux/xlog.h>
+#include <linux/types.h>
 
 #include "kd_camera_typedef.h"
 #include "kd_imgsensor.h"
@@ -1800,7 +1799,7 @@ static kal_uint32 feature_control(MSDK_SENSOR_FEATURE_ENUM feature_id,
 			*feature_para_len=4;
 			break;
 		case SENSOR_FEATURE_SET_FRAMERATE:
-            LOG_INF("current fps :%llu\n", *feature_data);
+            LOG_INF("current fps :%d\n", (UINT32)*feature_data);
 			spin_lock(&imgsensor_drv_lock);
             imgsensor.current_fps = *feature_data;
 			spin_unlock(&imgsensor_drv_lock);
@@ -1814,7 +1813,7 @@ static kal_uint32 feature_control(MSDK_SENSOR_FEATURE_ENUM feature_id,
 			spin_unlock(&imgsensor_drv_lock);
 			break;
 		case SENSOR_FEATURE_GET_CROP_INFO:
-            LOG_INF("SENSOR_FEATURE_GET_CROP_INFO scenarioId:%llu\n", *feature_data);
+            LOG_INF("SENSOR_FEATURE_GET_CROP_INFO scenarioId:%d\n", (UINT32)*feature_data);
             wininfo = (SENSOR_WINSIZE_INFO_STRUCT *)(uintptr_t)(*(feature_data+1));
 
 			switch (*feature_data_32) {
